chore: move test peer to test_utils
