+++ crates/p2p/src/test_utils.rs
@@ -1,3 +1,4 @@
 pub mod main_loop;
+pub mod peer;
 pub mod peer_aware;
 pub mod sync;
